fix(CatalogService): refactor & TODOs
