+++ packages/CatalogService/__tests__/utils/formatDataAsArray.ts
