+++ packages/CatalogService/__tests__/utils/makeStatic.ts
