+++ packages/CatalogService/apiFactory/operations/utils/printJSON.ts
